requirements.txt
azure-ai-projects>=2.0.0b3
azure-ai-agents
azure-identity
